pyproject.toml
[project]
name = "eap2-diffusion"
version = "0.1.0"
description = "Diffusion models for EAP2"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "accelerate>=1.4.0",
    "cupy-cuda12x>=13.3.0",
    "diffusers>=0.32.2",
    "imageio>=2.37.0",
    "ipykernel>=6.29.5",
    "jupytext>=1.16.7",
    "matplotlib>=3.9.4",
    "numba>=0.60.0",
    "numpy>=2.0.2",
    "pytest>=8.3.4",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
    "torch>=2.5.1",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
]



[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {  } # "QR1" = "models/QR1/", "QR2" = "models/QR2/"
packages = [  ] # ["QR1", "QR2"]



[tool.setuptools.package-data]
"*" = ["*.py", "*.pyi", "*.c", "*.h", "*.cpp", "*.hpp"]
